When OpenID plugin needs to create a new user, instead of assuming there's a "name" column in "User" model, it is better just reading all the arguments from CreateUser action.
The only column that is assumed have in user model is "openid", which
is defined in the model mixin.

Now the CreateOpenIDUser action is basically doing a CreateUser
action, but with some extra conditions. For example, the value of
openid must not be read from user input, but needs to be retrieved
from the session.


git-svn-id: svn+ssh://svn.bestpractical.com/svn/jifty.org/jifty/trunk@5536 e84bef0a-9b06-0410-84ba-c4c9edb13aeb